Réarrangement des fichiers HTML et rajout de delete pour list user

+++ src/main/java/com/mkyong/entity/User.java
@@ -17,7 +17,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Integer idUser;
+    private Long idUser;
 
     @Column(nullable=false,name="NOM_USER")
     @NotEmpty()
@@ -56,7 +56,7 @@ public class User {
 
     //getters
 
-    public Integer getIdUser() {
+    public Long getIdUser() {
         return idUser;
     }
 
@@ -80,7 +80,7 @@ public class User {
 
     //setters
 
-    public void setIdUser(Integer idUser) {
+    public void setIdUser(Long idUser) {
         this.idUser = idUser;
     }
